# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Order Hub API — multi-platform inventory and order management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: Paginated inventory list
      parameters:
        - in: query
          name: platform
          schema:
            type: string
            enum: [meesho, flipkart, amazon, all]
            default: all
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [productName, sku, quantity, price, purchaseDate]
            default: purchaseDate
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 5
            maximum: 100
            default: 20
        - in: query
          name: purchaseDateFrom
          description: Inclusive lower bound for purchaseDate (ISO 8601 date or datetime)
          schema:
            type: string
        - in: query
          name: purchaseDateTo
          description: Inclusive upper bound for purchaseDate (ISO 8601 date or datetime)
          schema:
            type: string
        - in: query
          name: allCompanies
          description: When true, returns inventory across every company instead of just the active brand
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Inventory page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryPage"

    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get a single inventory item by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory quantity / price / product name
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /inventory/bulk-delete:
    post:
      operationId: bulkDeleteInventory
      tags: [inventory]
      summary: Delete multiple inventory items by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeletePayload"
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Platform:
      type: string
      enum: [meesho, flipkart, amazon]
    Brand:
      type: string
      description: |
        Active brand workspace identifier (a company `slug`). Sent on every
        authenticated request via the `X-Brand` header. Companies are managed
        dynamically via the `/companies` endpoints.
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        brand:
          $ref: "#/components/schemas/Brand"
        platform:
          $ref: "#/components/schemas/Platform"
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        gstPercent:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        purchaseDate:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, brand, platform, productName, sku, quantity, price, updatedAt]
    InventoryInput:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/Platform"
        productName:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 0
        price:
          type: number
          minimum: 0
        originalPrice:
          type: number
          minimum: 0
        gstPercent:
          type: number
          minimum: 0
          maximum: 100
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        purchaseDate:
          type: string
          format: date-time
      required: [platform, productName, sku, quantity, price]
    InventoryUpdate:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/Platform"
        productName:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 0
        price:
          type: number
          minimum: 0
        originalPrice:
          type: number
          minimum: 0
        gstPercent:
          type: number
          minimum: 0
          maximum: 100
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        purchaseDate:
          type: string
          format: date-time
    BulkDeletePayload:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
      required: [ids]
    BulkDeleteResult:
      type: object
      properties:
        deleted:
          type: integer
      required: [deleted]
    InventoryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        totalCost:
          type: number
        totalUnits:
          type: integer
      required: [items, total, page, pageSize, totalPages, totalCost, totalUnits]
...